+++ .ci/generate.sh
@@ -8,20 +8,24 @@ if [ $? != 0 ]; then
     fi
 fi
 
+CONTROLLERGEN=controller-gen
+command -v ${CONTROLLERGEN} > /dev/null
+if [ $? != 0 ]; then
+    if [ -n ${GOPATH} ]; then
+        CONTROLLERGEN="${GOPATH}/bin/controller-gen"
+    fi
+fi
+
 # generate the CRD(s)
-operator-sdk generate crds
+${CONTROLLERGEN} crd paths=./pkg/apis/jaegertracing/... crd:maxDescLen=0,trivialVersions=true output:dir=./deploy/crds/
 RT=$?
 if [ ${RT} != 0 ]; then
     echo "Failed to generate CRDs."
     exit ${RT}
 fi
 
-# revert the APIs that are not managed by us
-for revert in deploy/crds/kafka.strimzi.io_kafkas_crd.yaml deploy/crds/kafka.strimzi.io_kafkausers_crd.yaml; do
-    if [ -f "${revert}" ]; then
-        rm "${revert}"
-    fi
-done
+# move the generated CRD to the same location the operator-sdk places
+mv deploy/crds/jaegertracing.io_jaegers.yaml deploy/crds/jaegertracing.io_jaegers_crd.yaml
 
 # generate the schema validation (openapi) stubs
 ${OPENAPIGEN} --logtostderr=true -o "" -i ./pkg/apis/jaegertracing/v1 -O zz_generated.openapi -p ./pkg/apis/jaegertracing/v1 -h /dev/null -r "-"