+++ pkg/controller/deployment/deployment_controller.go
@@ -3,13 +3,16 @@ package deployment
 import (
 	"context"
 
-	"k8s.io/apimachinery/pkg/types"
-
 	log "github.com/sirupsen/logrus"
+	"github.com/spf13/viper"
+	"go.opentelemetry.io/otel/api/key"
+	"go.opentelemetry.io/otel/global"
+	"google.golang.org/grpc/codes"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
@@ -19,6 +22,7 @@ import (
 
 	v1 "github.com/jaegertracing/jaeger-operator/pkg/apis/jaegertracing/v1"
 	"github.com/jaegertracing/jaeger-operator/pkg/inject"
+	"github.com/jaegertracing/jaeger-operator/pkg/tracing"
 )
 
 // Add creates a new Deployment Controller and adds it to the Manager. The Manager will set fields on the Controller
@@ -29,7 +33,7 @@ func Add(mgr manager.Manager) error {
 
 // newReconciler returns a new reconcile.Reconciler
 func newReconciler(mgr manager.Manager) reconcile.Reconciler {
-	return &ReconcileDeployment{client: mgr.GetClient(), scheme: mgr.GetScheme()}
+	return &ReconcileDeployment{client: mgr.GetClient(), scheme: mgr.GetScheme(), rClient: mgr.GetAPIReader()}
 }
 
 // add adds a new Controller to mgr with r as the reconcile.Reconciler
@@ -56,6 +60,11 @@ type ReconcileDeployment struct {
 	// This client, initialized using mgr.Client() above, is a split client
 	// that reads objects from the cache and writes to the apiserver
 	client client.Client
+
+	// this avoid the cache, which we need to bypass because the default client will attempt to place
+	// a watch on Namespace at cluster scope, which isn't desirable to us...
+	rClient client.Reader
+
 	scheme *runtime.Scheme
 }
 
@@ -67,6 +76,11 @@ type ReconcileDeployment struct {
 func (r *ReconcileDeployment) Reconcile(request reconcile.Request) (reconcile.Result, error) {
 	ctx := context.Background()
 
+	tracer := global.TraceProvider().GetTracer(v1.ReconciliationTracer)
+	ctx, span := tracer.Start(ctx, "reconcileDeployment")
+	defer span.End()
+
+	span.SetAttributes(key.String("name", request.Name), key.String("namespace", request.Namespace))
 	log.WithFields(log.Fields{
 		"namespace": request.Namespace,
 		"name":      request.Name,
@@ -80,32 +94,32 @@ func (r *ReconcileDeployment) Reconcile(request reconcile.Request) (reconcile.Re
 			// Request object not found, could have been deleted after reconcile request.
 			// Owned objects are automatically garbage collected. For additional cleanup logic use finalizers.
 			// Return and don't requeue
+			span.SetStatus(codes.NotFound)
 			return reconcile.Result{}, nil
 		}
 		// Error reading the object - requeue the request.
-		return reconcile.Result{}, err
+		return reconcile.Result{}, tracing.HandleError(err, span)
 	}
 
 	ns := &corev1.Namespace{}
-	err = r.client.Get(ctx, types.NamespacedName{Name: request.Namespace}, ns)
+	err = r.rClient.Get(ctx, types.NamespacedName{Name: request.Namespace}, ns)
+	// we shouldn't fail if the namespace object can't be obtained
 	if err != nil {
-		if errors.IsNotFound(err) {
-			// Request object not found, could have been deleted after reconcile request.
-			// Owned objects are automatically garbage collected. For additional cleanup logic use finalizers.
-			// Return and don't requeue
-			return reconcile.Result{}, nil
-		}
-		// Error reading the object - requeue the request.
-		return reconcile.Result{}, nil
+		log.WithField("namespace", request.Namespace).WithError(err).Trace("failed to get the namespace for the deployment, skipping injection based on namespace annotation")
+		tracing.HandleError(err, span)
 	}
 
 	if inject.Needed(dep, ns) {
 		jaegers := &v1.JaegerList{}
 		opts := []client.ListOption{}
-		err := r.client.List(ctx, jaegers, opts...)
-		if err != nil {
+
+		if viper.GetString(v1.ConfigOperatorScope) == v1.OperatorScopeNamespace {
+			opts = append(opts, client.InNamespace(viper.GetString(v1.ConfigWatchNamespace)))
+		}
+
+		if err := r.rClient.List(ctx, jaegers, opts...); err != nil {
 			log.WithError(err).Error("failed to get the available Jaeger pods")
-			return reconcile.Result{}, err
+			return reconcile.Result{}, tracing.HandleError(err, span)
 		}
 
 		jaeger := inject.Select(dep, ns, jaegers)
@@ -121,7 +135,7 @@ func (r *ReconcileDeployment) Reconcile(request reconcile.Request) (reconcile.Re
 			dep = inject.Sidecar(jaeger, dep)
 			if err := r.client.Update(ctx, dep); err != nil {
 				log.WithField("deployment", dep).WithError(err).Error("failed to update")
-				return reconcile.Result{}, err
+				return reconcile.Result{}, tracing.HandleError(err, span)
 			}
 		} else {
 			log.WithField("deployment", dep.Name).Info("No suitable Jaeger instances found to inject a sidecar")