+++ pkg/controller/namespace/namespace_controller.go
@@ -4,6 +4,10 @@ import (
 	"context"
 
 	log "github.com/sirupsen/logrus"
+	"github.com/spf13/viper"
+	"go.opentelemetry.io/otel/api/key"
+	"go.opentelemetry.io/otel/global"
+	"google.golang.org/grpc/codes"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -17,6 +21,7 @@ import (
 
 	v1 "github.com/jaegertracing/jaeger-operator/pkg/apis/jaegertracing/v1"
 	"github.com/jaegertracing/jaeger-operator/pkg/inject"
+	"github.com/jaegertracing/jaeger-operator/pkg/tracing"
 )
 
 // Add creates a new Namespace Controller and adds it to the Manager. The Manager will set fields on the Controller
@@ -27,11 +32,17 @@ func Add(mgr manager.Manager) error {
 
 // newReconciler returns a new reconcile.Reconciler
 func newReconciler(mgr manager.Manager) reconcile.Reconciler {
-	return &ReconcileNamespace{client: mgr.GetClient(), scheme: mgr.GetScheme()}
+	return &ReconcileNamespace{client: mgr.GetClient(), scheme: mgr.GetScheme(), rClient: mgr.GetAPIReader()}
 }
 
 // add adds a new Controller to mgr with r as the reconcile.Reconciler
 func add(mgr manager.Manager, r reconcile.Reconciler) error {
+	// we only create this controller if we have cluster-scope, as watching namespaces is a cluster-wide operation
+	if !viper.GetBool(v1.ConfigEnableNamespaceController) {
+		log.Trace("skipping reconciliation for namespaces, do not have permissions to list and watch namespaces")
+		return nil
+	}
+
 	// Create a new controller
 	c, err := controller.New("namespace-controller", mgr, controller.Options{Reconciler: r})
 	if err != nil {
@@ -53,6 +64,11 @@ type ReconcileNamespace struct {
 	// This client, initialized using mgr.Client() above, is a split client
 	// that reads objects from the cache and writes to the apiserver
 	client client.Client
+
+	// this avoid the cache, which we need to bypass because the default client will attempt to place
+	// a watch on Namespace at cluster scope, which isn't desirable to us...
+	rClient client.Reader
+
 	scheme *runtime.Scheme
 }
 
@@ -64,22 +80,28 @@ type ReconcileNamespace struct {
 func (r *ReconcileNamespace) Reconcile(request reconcile.Request) (reconcile.Result, error) {
 	ctx := context.Background()
 
+	tracer := global.TraceProvider().GetTracer(v1.ReconciliationTracer)
+	ctx, span := tracer.Start(ctx, "reconcileNamespace")
+	defer span.End()
+
+	span.SetAttributes(key.String("name", request.Name), key.String("namespace", request.Namespace))
 	log.WithFields(log.Fields{
 		"namespace": request.Namespace,
 		"name":      request.Name,
 	}).Trace("Reconciling Namespace")
 
 	ns := &corev1.Namespace{}
-	err := r.client.Get(ctx, request.NamespacedName, ns)
+	err := r.rClient.Get(ctx, request.NamespacedName, ns)
 	if err != nil {
 		if errors.IsNotFound(err) {
 			// Request object not found, could have been deleted after reconcile request.
 			// Owned objects are automatically garbage collected. For additional cleanup logic use finalizers.
 			// Return and don't requeue
+			span.SetStatus(codes.NotFound)
 			return reconcile.Result{}, nil
 		}
 		// Error reading the object - requeue the request.
-		return reconcile.Result{}, nil
+		return reconcile.Result{}, tracing.HandleError(err, span)
 	}
 
 	opts := []client.ListOption{
@@ -88,7 +110,7 @@ func (r *ReconcileNamespace) Reconcile(request reconcile.Request) (reconcile.Res
 
 	// Fetch the Deployment instance
 	deps := &appsv1.DeploymentList{}
-	err = r.client.List(ctx, deps, opts...)
+	err = r.rClient.List(ctx, deps, opts...)
 	if err != nil {
 		if errors.IsNotFound(err) {
 			// Request object not found, could have been deleted after reconcile request.
@@ -97,7 +119,7 @@ func (r *ReconcileNamespace) Reconcile(request reconcile.Request) (reconcile.Res
 			return reconcile.Result{}, nil
 		}
 		// Error reading the object - requeue the request.
-		return reconcile.Result{}, err
+		return reconcile.Result{}, tracing.HandleError(err, span)
 	}
 
 	for i := 0; i < len(deps.Items); i++ {
@@ -105,10 +127,14 @@ func (r *ReconcileNamespace) Reconcile(request reconcile.Request) (reconcile.Res
 		if inject.Needed(dep, ns) {
 			jaegers := &v1.JaegerList{}
 			opts := []client.ListOption{}
-			err := r.client.List(ctx, jaegers, opts...)
-			if err != nil {
+
+			if viper.GetString(v1.ConfigOperatorScope) == v1.OperatorScopeNamespace {
+				opts = append(opts, client.InNamespace(viper.GetString(v1.ConfigWatchNamespace)))
+			}
+
+			if err := r.rClient.List(ctx, jaegers, opts...); err != nil {
 				log.WithError(err).Error("failed to get the available Jaeger pods")
-				return reconcile.Result{}, err
+				return reconcile.Result{}, tracing.HandleError(err, span)
 			}
 
 			jaeger := inject.Select(dep, ns, jaegers)
@@ -124,7 +150,7 @@ func (r *ReconcileNamespace) Reconcile(request reconcile.Request) (reconcile.Res
 				dep = inject.Sidecar(jaeger, dep)
 				if err := r.client.Update(ctx, dep); err != nil {
 					log.WithField("deployment", dep).WithError(err).Error("failed to update")
-					return reconcile.Result{}, err
+					return reconcile.Result{}, tracing.HandleError(err, span)
 				}
 			} else {
 				log.WithField("deployment", dep.Name).Info("No suitable Jaeger instances found to inject a sidecar")